Implementaci[on de comunicacion spi y adc

+++ lab3/Lab3/ADC/ADC.h
@@ -11,8 +11,8 @@
 #include <avr/io.h>
 #include <util/delay.h>
 
-extern volatile uint16_t adc3;
-extern volatile uint16_t adc4;
+extern volatile uint8_t adc3;
+extern volatile uint8_t adc4;
 
 void init_ADC(void);
